# openapispec.yaml
openapi: 3.1.0
...
info:
  title: melodyMingle
  version: '1.0'
servers:
  - url: 'http://localhost:3000'
paths:
  '/users/{userId}':
    parameters:
      - schema:
          type: string
        name: userId
        in: path
        required: true
    get:
      summary: Get User Info by User ID
      tags: []
      responses:
        '200':
          description: User Found
          content: {}
        '404':
          description: User Not Found
      operationId: get-users-userId
      description: Retrieve the information of the user with the matching user ID.
      x-stoplight:
        id: bcckyc996fsg8
      parameters:
        - schema:
            type: string
          in: header
          name: Authorization
          description: access token
    delete:
      summary: Delete User by User ID
      tags: []
      responses: {}
      operationId: delete-users-userId
      x-stoplight:
        id: suam72d59n4gx
      parameters:
        - schema:
            type: string
          in: header
          name: Authorization
          description: access token
    put:
      summary: Update User by User ID
      tags: []
      responses: {}
      operationId: put-users-userId
      x-stoplight:
        id: 8n3sx3yu1e3i6
      parameters:
        - schema:
            type: string
          in: header
          name: Authorization
          description: access token
...
  '/groups/{groupId}':
    parameters:
      - schema:
          type: string
        name: groupId
        in: path
        required: true
    get:
      summary: Get a specific group by ID
      tags: []
      responses: {}
      operationId: get-groups-groupId
      x-stoplight:
        id: 62kylg3h3i5fa
    delete:
      summary: Delete a group by ID
      tags: []
      responses: {}
      operationId: delete-groups-groupId
      x-stoplight:
        id: h7s8my6e5bnqg
    put:
      summary: Update a specific group
      tags: []
      responses: {}
      operationId: put-groups-groupId
      x-stoplight:
        id: w52vx8az1ttsn
...
  '/users/upload/{userId}':
    parameters:
      - schema:
          type: string
        name: userId
        in: path
        required: true
      - schema:
          type: string
        name: userId
        in: path
        required: true
    post:
      summary: Upload profile picture for user
      tags: []
      responses: {}
      operationId: post-users-upload-userId
      x-stoplight:
        id: x99qo63iztass
...